move query/signal handlers into a separate file
