+++ apps/hash-ai-worker-ts/src/workflows/run-flow-workflow.ts
@@ -1,17 +1,13 @@
-import type { SentrySinks } from "@local/hash-backend-utils/temporal/sinks/sentry";
 import { sleep } from "@local/hash-backend-utils/utils";
 import {
   type ActionDefinitionId,
   actionDefinitions,
 } from "@local/hash-isomorphic-utils/flows/action-definitions";
-import { externalInputResponseSignal } from "@local/hash-isomorphic-utils/flows/signals";
 import type {
   RunFlowWorkflowParams,
   RunFlowWorkflowResponse,
 } from "@local/hash-isomorphic-utils/flows/temporal-types";
 import type {
-  ExternalInputRequestSignal,
-  ExternalInputResponseSignal,
   FlowDefinition,
   FlowStep,
   Payload,
@@ -24,8 +20,6 @@ import { StatusCode } from "@local/status";
 import {
   ActivityCancellationType,
   proxyActivities,
-  proxySinks,
-  setHandler,
   workflowInfo,
 } from "@temporalio/workflow";
 
@@ -34,8 +28,6 @@ import type {
   createFlowActivities,
 } from "../activities/flow-activities";
 import { stringify } from "../activities/shared/stringify";
-import { getExternalInputResponseQuery } from "../shared/queries";
-import { externalInputRequestSignal } from "../shared/signals";
 import { getAllStepsInFlow } from "./run-flow-workflow/get-all-steps-in-flow";
 import { getStepDefinitionFromFlowDefinition } from "./run-flow-workflow/get-step-definition-from-flow";
 import {
@@ -44,13 +36,12 @@ import {
   initializeParallelGroup,
 } from "./run-flow-workflow/initialize-flow";
 import { passOutputsToUnprocessedSteps } from "./run-flow-workflow/pass-outputs-to-unprocessed-steps";
+import { setQueryAndSignalHandlers } from "./run-flow-workflow/set-query-and-signal-handlers";
 
 const log = (message: string) => {
   // eslint-disable-next-line no-console
   console.log(message);
 };
-
-const sinks = proxySinks<SentrySinks>();
 
 const doesFlowStepHaveSatisfiedDependencies = (params: {
   step: FlowStep;
@@ -206,80 +197,7 @@ export const runFlowWorkflow = async (
 
   await flowActivities.persistFlowActivity({ flow, userAuthentication });
 
-  /**
-   * Handle requests from activities for external input (e.g. human input, HTML from a browser authenticated with a
-   * site).
-   *
-   * The process by which these requests and responses occur is:
-   * 1. Activity sends a Signal to the Workflow requesting data from the outside world when it encounters a need for it
-   * 2. The outside world is polling for these requests (resolved by finding the signals in the event history, for now)
-   * 3. The outside world sends a Signal back to the Workflow with the requested data
-   * 4. The Workflow allows the response to be accessed via a Query
-   */
-  const externalInputRequestsById = new Map<
-    string,
-    {
-      request: ExternalInputRequestSignal;
-      response?: ExternalInputResponseSignal;
-    }
-  >();
-
-  setHandler(
-    externalInputRequestSignal,
-    (request: ExternalInputRequestSignal) => {
-      if (!externalInputRequestsById.has(request.requestId)) {
-        externalInputRequestsById.set(request.requestId, { request });
-      }
-    },
-  );
-
-  setHandler(
-    externalInputResponseSignal,
-    (response: ExternalInputResponseSignal) => {
-      const { requestId } = response;
-      const inputRequestRecord = externalInputRequestsById.get(requestId);
-
-      if (!inputRequestRecord) {
-        /**
-         * It's not clear in what circumstances this will happen, and we can't do much about it,
-         * but we should log it to be aware if it is happening.
-         */
-        sinks.sentry.captureException(
-          new Error(
-            `Received response for external input request ${requestId}, but no record of request was found`,
-          ),
-        );
-        return;
-      }
-
-      if (response.type !== inputRequestRecord.request.type) {
-        sinks.sentry.captureException(
-          new Error(
-            `Response for external input request ${requestId} has type ${response.type}, but expected ${inputRequestRecord.request.type}`,
-          ),
-        );
-        return;
-      }
-
-      inputRequestRecord.response = response;
-    },
-  );
-
-  setHandler(getExternalInputResponseQuery, ({ requestId }) => {
-    const inputRequestRecord = externalInputRequestsById.get(requestId);
-    if (!inputRequestRecord) {
-      /**
-       * We can throw an error here to crash the activity, since it is waiting for a response it will never receive.
-       *
-       */
-      throw new Error(
-        `Received query for response to external input request ${requestId}, but no record of request was found`,
-      );
-    }
-
-    return inputRequestRecord.response ?? null;
-  });
-  /** End of external input handling */
+  setQueryAndSignalHandlers();
 
   const processedStepIds: string[] = [];
   const processStepErrors: Record<string, Omit<Status<never>, "contents">> = {};
